Quick Bug Fix and Extra Defensive Lines
There was a bug about primery keys in db, so i fixed that by adding extra id column in entities. Also I added defense in AddPost-CommentLike repos. It now checks db to retrieve data. If its in db, it's not going to add extra entity to db, it returns existing entity.